+++ src/util/Math.h
@@ -3,8 +3,7 @@
 
 namespace stellar
 {
-float rand_fraction();
+double rand_fraction();
 
 size_t rand_pareto(float alpha, size_t max);
-
 }
